Tidying up the template
The template this project was based on had inconsistency between tabs and
spaces. I've standardized on spaces.

And it used a relative path in an #include statement - which is fragile. This
has been removed by adding a 'target_include_directories' CMake directive to
help lcc find the project's include files.