Import and basic integration with cbt-fx sound effects library
This needed a patch as cbt-fx calls into the hUGEDriver music
library. hUGEDriver has an asm function for muting and unmuting sound channels
in the music while a sound effect plays, but this asm function was written to
the new calling convention that sdcc 4.1 and later use, whereas cbt-fx has an
OLDCALL keyword on its hUGEDriver function prototypes.

A player selection sound was implemented and is played when the player selection
cursor is used to select a bot in the title screen.