Add joypad reading to the level's update loop
Problem
=======

The main level will need to read the joypad. But it is initialized in the
title screen.

Solution
========

Factors out joypad initialization up into main(), and passes in the joypad
struct into both the title screen and the main level.

Testing
=======

Ran in 'bgb' emulator. The title screen was able to choose robots still.