Let the joypad control the bot in the level.
Problem
=======

The bot ought be able to move in response to joypad input, but it does not.

Solution
========

Add 'update_velocities' and 'update_positions' functions in the main level
update loop that uses joypad input to set the bot entity's velocity, and to
update the bot entity position based on velocity.

Testing
=======

Ran this in 'bgb'. The bot can freely move around the screen now in response to
joypad input. It's not constrained to the level grid yet however.